+++ httprunner/utils_test.py
@@ -86,75 +86,6 @@ class TestUtils(unittest.TestCase):
         new_request_dict = utils.lower_dict_keys(request_dict)
         self.assertEqual(None, request_dict)
 
-    def test_deepcopy_dict(self):
-        license_path = os.path.join(
-            os.path.dirname(os.path.dirname(__file__)), "LICENSE"
-        )
-        data = {
-            "a": 1,
-            "b": [2, 4],
-            "c": lambda x: x + 1,
-            "d": open(license_path),
-            "f": {"f1": {"a1": 2}, "f2": io.open(license_path, "rb"),},
-        }
-        new_data = utils.deepcopy_dict(data)
-        data["a"] = 0
-        self.assertEqual(new_data["a"], 1)
-        data["f"]["f1"] = 123
-        self.assertEqual(new_data["f"]["f1"], {"a1": 2})
-        self.assertNotEqual(id(new_data["b"]), id(data["b"]))
-        self.assertEqual(id(new_data["c"]), id(data["c"]))
-        # self.assertEqual(id(new_data["d"]), id(data["d"]))
-
-    def test_cartesian_product_one(self):
-        parameters_content_list = [[{"a": 1}, {"a": 2}]]
-        product_list = utils.gen_cartesian_product(*parameters_content_list)
-        self.assertEqual(product_list, [{"a": 1}, {"a": 2}])
-
-    def test_cartesian_product_multiple(self):
-        parameters_content_list = [
-            [{"a": 1}, {"a": 2}],
-            [{"x": 111, "y": 112}, {"x": 121, "y": 122}],
-        ]
-        product_list = utils.gen_cartesian_product(*parameters_content_list)
-        self.assertEqual(
-            product_list,
-            [
-                {"a": 1, "x": 111, "y": 112},
-                {"a": 1, "x": 121, "y": 122},
-                {"a": 2, "x": 111, "y": 112},
-                {"a": 2, "x": 121, "y": 122},
-            ],
-        )
-
-    def test_cartesian_product_empty(self):
-        parameters_content_list = []
-        product_list = utils.gen_cartesian_product(*parameters_content_list)
-        self.assertEqual(product_list, [])
-
     def test_print_info(self):
         info_mapping = {"a": 1, "t": (1, 2), "b": {"b1": 123}, "c": None, "d": [4, 5]}
         utils.print_info(info_mapping)
-
-    def test_prepare_dump_json_file_path_for_folder(self):
-        # hrun tests/httpbin/a.b.c/ --save-tests
-        test_path = os.path.join("tests", "httpbin", "a.b.c")
-        self.assertEqual(
-            utils.prepare_log_file_abs_path(test_path, "loaded.json"),
-            os.path.join(os.getcwd(), "logs", "tests/httpbin/a.b.c/all.loaded.json"),
-        )
-
-    def test_prepare_dump_json_file_path_for_file(self):
-        # hrun tests/httpbin/a.b.c/rpc.yml --save-tests
-        test_path = os.path.join("tests", "httpbin", "a.b.c", "rpc.yml")
-        self.assertEqual(
-            utils.prepare_log_file_abs_path(test_path, "loaded.json"),
-            os.path.join(os.getcwd(), "logs", "tests/httpbin/a.b.c/rpc.loaded.json"),
-        )
-
-    def test_prepare_dump_json_file_path_for_passed_testcase(self):
-        test_path = ""
-        self.assertEqual(
-            utils.prepare_log_file_abs_path(test_path, "loaded.json"),
-            os.path.join(os.getcwd(), "logs", "tests_mapping.loaded.json"),
-        )
